+++ packages/@expo/cli/e2e/playwright/dev/fast-refresh.test.ts
@@ -28,7 +28,6 @@ test.describe(inputDir, () => {
       NODE_ENV: 'development',
       EXPO_USE_STATIC: 'single',
       E2E_ROUTER_SRC: inputDir,
-      E2E_ROUTER_ASYNC: 'development',
 
       // Ensure CI is disabled otherwise the file watcher won't run.
       CI: '0',